refactor: extract CLI input-resolution cluster into cli_resolve leaf (C7-tail step 1)
Move the input-resolution & native-dump dispatch cluster out of cli.py into a
new leaf module abicheck/cli_resolve.py:

  _expand_header_inputs, _sniff_text_format, _detect_binary_format,
  _resolve_linker_script, _maybe_follow_linker_script, _normalize_binary_input,
  _dump_elf, _apply_native_provenance, _dump_native_binary, _resolve_input,
  _populate_dependency_info, _is_supported_compare_input, _resolve_compare_snapshots

cli_resolve depends only on lower-level modules (dumper, serialization, errors,
compat, and lazily binary_utils/service/provenance/binder/resolver/classify), so
the dependency runs one way (cli -> cli_resolve) and stays cycle-free.

The helpers are re-exported from cli (declared in __all__ so the re-export is
explicit for mypy's no-implicit-reexport and for ruff), keeping every existing
`from abicheck.cli import _resolve_input` call site — sibling cli_* modules,
mcp_server, and the test suite — working unchanged.

Also drop the dead `_LD_SCRIPT_RE` / `_LD_KEYWORDS` regexes from cli (the live
copies live in binary_utils, which _resolve_linker_script already delegates to).

cli.py drops from 1928 to 1519 lines (under the 2000-line hard cap). Stale
monkeypatch targets that referenced functions in their old home are repointed to
abicheck.cli_resolve (intent unchanged; the dump-command tests that patch via
cli's namespace are intentionally left as-is). Full fast suite green.
